requirements.txt
openai==0.27.2
google-cloud-texttospeech==2.14.1
python-dotenv==0.21.0
